+++ contracts/libraries/OracleLib.sol
@@ -10,8 +10,9 @@ library OracleLib {
 
     struct Observation {
         uint32 blockTimestamp;
-        uint128 lnImpliedRateCumulative;
+        uint216 lnImpliedRateCumulative;
         bool initialized;
+        // 1 SLOT = 256 bits
     }
 
     struct OracleData {
@@ -30,7 +31,7 @@ library OracleLib {
             Observation({
                 blockTimestamp: blockTimestamp,
                 lnImpliedRateCumulative: last.lnImpliedRateCumulative +
-                    uint128(lnImpliedRate) *
+                    uint216(lnImpliedRate) *
                     (blockTimestamp - last.blockTimestamp),
                 initialized: true
             });
@@ -166,7 +167,7 @@ library OracleLib {
         uint96 lnImpliedRate,
         uint16 index,
         uint16 cardinality
-    ) private view returns (uint128 lnImpliedRateCumulative) {
+    ) private view returns (uint216 lnImpliedRateCumulative) {
         if (secondsAgo == 0) {
             Observation memory last = obv[index];
             if (last.blockTimestamp != time) {
@@ -194,7 +195,7 @@ library OracleLib {
         } else {
             // we're in the middle
             return (beforeOrAt.lnImpliedRateCumulative +
-                uint128(
+                uint216(
                     (uint256(
                         atOrAfter.lnImpliedRateCumulative - beforeOrAt.lnImpliedRateCumulative
                     ) * (target - beforeOrAt.blockTimestamp)) /
@@ -208,10 +209,10 @@ library OracleLib {
         uint32 time,
         uint32[] memory secondsAgos,
         uint96 lnImpliedRate
-    ) public view returns (uint128[] memory lnImpliedRateCumulative) {
+    ) public view returns (uint216[] memory lnImpliedRateCumulative) {
         require(self.cardinality != 0, "cardinality must be positive");
 
-        lnImpliedRateCumulative = new uint128[](secondsAgos.length);
+        lnImpliedRateCumulative = new uint216[](secondsAgos.length);
         for (uint256 i = 0; i < lnImpliedRateCumulative.length; ) {
             lnImpliedRateCumulative[i] = observeSingle(
                 self.observations,
@@ -239,7 +240,7 @@ library OracleLib {
         secondsAgos[0] = secondsAgo;
         secondsAgos[1] = 0;
 
-        uint128[] memory lnImpliedRateCumulatives = observe(
+        uint216[] memory lnImpliedRateCumulatives = observe(
             self,
             time,
             secondsAgos,
